% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fitBMWithShiftsToMammalData}
\alias{fitBMWithShiftsToMammalData}
\title{An example model fit of BM model with shifts to a large phylogenetic tree
with bivariate trait data.}
\format{This is a list with assigned S3 class `PCMFitModelMappings`.}
\usage{
fitBMWithShiftsToMammalData
}
\description{
An example model fit of BM model with shifts to a large phylogenetic tree
with bivariate trait data.
}
\keyword{datasets}
